test: consolidate hook executor tests into single file
Issue: Test files for hook executor were split unnecessarily into two separate files
- hookExecutor.test.ts contained tests for executeHook and executeWorktreePostCreationHook
- hookExecutor.statusHook.test.ts contained tests for executeStatusHook
- This separation made test maintenance more difficult

Reason: Having multiple test files for a single module adds unnecessary complexity
- Tests for the same module should be colocated for better organization
- Separate files increase cognitive load when understanding test coverage
- Duplication of imports and mock setup across files

Solution: Merged hookExecutor.statusHook.test.ts into hookExecutor.test.ts
- Moved all executeStatusHook tests into main test file
- Consolidated mock setup at the top of the file
- Maintained all existing test coverage and functionality
- Deleted the now-redundant statusHook test file
